Correctly handle hostnames with and without trailing dot in the DefaultDnsCache.
Motivation:

We should ensure we return the same cached entries for the hostname and hostname ending with dot.

Modifications:

Internally always use hostname with a dot as a key and so ensure we correctly handle it in the cache.

Result:

Use the same cached entries for hostname with and without trailing dot.